+++ src/gtirb-decoder/target/RawMips32Loader.h
@@ -1,6 +1,6 @@
-//===- PeX64Loader.h --------------------------------------------*- C++ -*-===//
+//===- RawMips32Loader.h ----------------------------------------*- C++ -*-===//
 //
-//  Copyright (C) 2020 GrammaTech, Inc.
+//  Copyright (C) 2021 GrammaTech, Inc.
 //
 //  This code is licensed under the GNU Affero General Public License
 //  as published by the Free Software Foundation, either version 3 of
@@ -20,26 +20,33 @@
 //  endorsement should be inferred.
 //
 //===----------------------------------------------------------------------===//
-#ifndef SRC_GTIRB_DECODER_TARGET_PEX64LOADER_H_
-#define SRC_GTIRB_DECODER_TARGET_PEX64LOADER_H_
+#ifndef SRC_GTIRB_DECODER_TARGET_RAWMIPS32LOADER_H_
+#define SRC_GTIRB_DECODER_TARGET_RAWMIPS32LOADER_H_
 
 #include "../CompositeLoader.h"
-#include "../arch/X64Loader.h"
 #include "../core/DataLoader.h"
-#include "../core/ModuleLoader.h"
-#include "../core/SectionLoader.h"
-#include "../format/PeLoader.h"
+#include "../format/RawLoader.h"
 
-CompositeLoader PeX64Loader()
+CompositeLoader RawMips32BELoader()
 {
-    CompositeLoader Loader("souffle_disasm_x86_64");
+    CompositeLoader Loader("souffle_disasm_mips32");
     Loader.add(ModuleLoader);
     Loader.add(SectionLoader);
-    Loader.add<X64Loader>();
-    Loader.add<DataLoader>(DataLoader::Pointer::QWORD);
-    Loader.add(PeSymbolLoader);
-    Loader.add(PeDataDirectoryLoader);
+    Loader.add<Mips32Loader>(Mips32Loader::Endianness::BIG);
+    Loader.add<DataLoader>(DataLoader::Pointer::DWORD, DataLoader::Endianness::BIG);
+    Loader.add(RawEntryLoader);
     return Loader;
-};
+}
 
-#endif // SRC_GTIRB_DECODER_TARGET_PEX64LOADER_H_
+CompositeLoader RawMips32LELoader()
+{
+    CompositeLoader Loader("souffle_disasm_mips32");
+    Loader.add(ModuleLoader);
+    Loader.add(SectionLoader);
+    Loader.add<Mips32Loader>(Mips32Loader::Endianness::LITTLE);
+    Loader.add<DataLoader>(DataLoader::Pointer::DWORD, DataLoader::Endianness::LITTLE);
+    Loader.add(RawEntryLoader);
+    return Loader;
+}
+
+#endif // SRC_GTIRB_DECODER_TARGET_MIPS32LOADER_H_