+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.4.1
+
+* Use dedicated symbolic expression attributes.
+* Support Souffle interpreter for development.
+* Add support for loading existing GTIRB files.
+
 # 1.4.0
 
 * Add PE support.